Move auth actions into a new folder & rename it
