+++ lib/actions/auth.js
@@ -2,7 +2,7 @@
 
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
-import { createClient as createServerClient } from '../lib/supabase/server';
+import { createClient as createServerClient } from '../supabase/server';
 
 //Authenticates & logs in user with email & password
 export async function signIn(formData) {
